requirements.txt
langchain>=0.0.267
langchain_community>=0.0.1
langchain_core>=0.0.1
langchain_experimental>=0.0.1
langchain-openai>=0.0.1
langgraph>=0.0.1
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
openai>=1.1.1
sqlalchemy>=2.0.0
pandas>=2.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-otlp>=1.18.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
